requirements.txt
numpy~=1.26.4
tensorflow~=2.10.0
matplotlib~=3.8.0
pillow~=10.2.0
scikit-learn~=1.3.0